requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-google-genai
langchain-core
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
flask
pandas
numpy
tqdm
ipython